Finished the slider with the functionality.
Brought the slider from the admin panel image drop off and modified it a little bit for this modal and guest/user permissions (no deleting the picture possible).
