+++ client/src/components/ProductDetailsCard/ProductDetailsPhotoSlider/ImageWrapper/styles.js
@@ -1,0 +1,14 @@
+import {
+    makeStyles
+} from '@material-ui/core/styles';
+
+const useStyles = makeStyles((theme) => ({
+    imageDis: {
+        display: "flex",
+        flexDirection: "column",
+        position: 'relative',
+        justify: 'space-between',
+    }
+}));
+
+export default useStyles;
